feat: Add workout type distribution chart to dashboard
Implemented a Doughnut chart to display workout type distribution using Chart.js. Adjusted chart positioning and styling with Tailwind CSS for a responsive layout. Initialized chart data with default values of 1 for a cleaner user experience.
Changed the id for the summarized workouts to match the workout day so we can render the edit workout page by matching ids